moving chevrons to /components
